VBPubSub
re: issue #22

Adds a basic pubsub mechanism in VBPubSub that allows listening to one channel and publishing to any channel.  It works by polling a "PubSub" class on Parse. It essentially simulates what you'd expect from a real-time pubsub system, except with polling changes won't be pushed instantly.

Also added a simple mechanism for listening to a user in VBInputSourceManager.

For testing, VBInputSourceManager publishes everything to the server, and listens from the server before sending out notifications. Since it works for an "anon" user and uses the same mechanism as listening to another user, it should work for listening to any user broadcasting from their microphone.

At the moment, everyone is the "same" anon user until we get the [user currentUser] pattern back. This means if we use the app at the same time, we should read what everyone is saying to the app!