Publish On Checkin
Currently, the authorized user is publishing to his/her channel. It's a waste of bandwidth given that nobody will listen until the user is checked in. This patch adds an "isCheckedIn" property and implements it in the source controller. Only if a user is checked into a venue will he/she start to publish to his/her respective channel.