Tap to subscribe to camera images
Closes #99

- User cell tweaks to show live updating camera feed
- Allow tapping on camera view in user cell to subscribe to new camera
- Animation of new camera view into the caption view

The way I've done it is maybe a little convulated, as I needed to propagate the animation to the root view controller. The VBUserTableViewCell fires a VBEventCameraSourceChanged with the new user for the camera source, the frame for an animation start placeholder, and the image from that camera source. The caption controller listens for any camera source changes, and tells it's root Controller to perform the animation and sets its new camera source image on completion.